requirement.txt
click==7.1.2
Flask==1.1.2
Flask-SQLAlchemy==2.4.3
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
psycopg2==2.8.5
pypng==0.0.20
PyQRCode==1.2.1
SQLAlchemy==1.3.17
Werkzeug==1.0.1
